python/requirements.txt
grpcio==1.63.0
grpcio-tools==1.63.0
protobuf==5.26.1
pytest==8.2.0